Use correct types to avoid sim warnings.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@7627 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/a52.c
@@ -122,7 +122,7 @@ extern char iramend[];
 /* this is the codec entry point */
 enum codec_status codec_start(struct codec_api *api)
 {
-    size_t n;
+    long n;
     unsigned char *filebuf;
 
     /* Generic codec initialisation */
